22-06-12 23:46

+++ Baekjoon/src/class1/B1008.java
@@ -6,8 +6,8 @@ public class B1008 {
 
 	public static void main(String[] args) {
 		Scanner sc = new Scanner(System.in);
-		int a = sc.nextInt();
-		int b = sc.nextInt();
+		double a = sc.nextInt();
+		double b = sc.nextInt();
 		System.out.println(a/b);
 
 	}
